cufree: fix warning on integer type
/home/fluids/hoefling/projects/gpgpu/nvcuda-tools/cufree/cufree.c: In function ‘cufree’:
/home/fluids/hoefling/projects/gpgpu/nvcuda-tools/cufree/cufree.c:40: warning: passing argument 1 of ‘cuMemGetInfo_v2’ from incompatible pointer type
/usr/local/cuda/include/cuda.h:2197: note: expected ‘size_t *’ but argument is of type ‘unsigned int *’
/home/fluids/hoefling/projects/gpgpu/nvcuda-tools/cufree/cufree.c:40: warning: passing argument 2 of ‘cuMemGetInfo_v2’ from incompatible pointer type
/usr/local/cuda/include/cuda.h:2197: note: expected ‘size_t *’ but argument is of type ‘unsigned int *’

+++ cufree/cufree.c
@@ -25,7 +25,7 @@ int cufree(int d)
     CUcontext ctx;
     CUdevice dev;
     CUresult err;
-    unsigned int free = 0, total = 0;
+    size_t free = 0, total = 0;
 
     if (CUDA_SUCCESS != cuDeviceGet(&dev, d)) {
 	fprintf(stderr, "cufree: failed to initialize the CUDA driver API\n", d);
